Refactored repost & derepost out of handlers.ts
ref https://linear.app/ghost/issue/PROD-669

In order to add restrictions on how can perform actions we need to pull the API
logic out into services, the first step of that is breaking out the handlers
into their own files.
