+++ src/handlers.ts
@@ -1,17 +1,7 @@
-import { createHash } from 'node:crypto';
-import { type Actor, Announce, PUBLIC_COLLECTION, Undo } from '@fedify/fedify';
 import type { Context } from 'hono';
 
-import { exhaustiveCheck, getError, getValue, isError } from 'core/result';
-import { parseURL } from 'core/url';
-import type { KnexAccountRepository } from './account/account.repository.knex';
 import { type HonoContextVariables, fedify } from './app';
-import { ACTOR_DEFAULT_HANDLE } from './constants';
 import { buildActivity } from './helpers/activitypub/activity';
-import { addToList, removeFromList } from './kv-helpers';
-import { lookupActor, lookupObject } from './lookup-helpers';
-import type { KnexPostRepository } from './post/post.repository.knex';
-import type { PostService } from './post/post.service';
 import type { SiteService } from './site/site.service';
 
 export const getSiteDataHandler =
@@ -112,259 +102,4 @@ export async function inboxHandler(
             status: 200,
         },
     );
-}
-
-export function createRepostActionHandler(
-    accountRepository: KnexAccountRepository,
-    postService: PostService,
-    postRepository: KnexPostRepository,
-) {
-    return async function repostAction(
-        ctx: Context<{ Variables: HonoContextVariables }>,
-    ) {
-        const id = ctx.req.param('id');
-        const apCtx = fedify.createContext(ctx.req.raw as Request, {
-            db: ctx.get('db'),
-            globaldb: ctx.get('globaldb'),
-            logger: ctx.get('logger'),
-        });
-
-        const post = await lookupObject(apCtx, id);
-        if (!post) {
-            return new Response(JSON.stringify({ error: 'Post not found' }), {
-                status: 404,
-            });
-        }
-
-        const announceId = apCtx.getObjectUri(Announce, {
-            id: createHash('sha256').update(post.id!.href).digest('hex'),
-        });
-
-        if (await ctx.get('globaldb').get([announceId.href])) {
-            return new Response(
-                JSON.stringify({ error: 'Post already reposted' }),
-                {
-                    status: 409,
-                },
-            );
-        }
-
-        await post.getAttribution();
-
-        const actor = await apCtx.getActor(ACTOR_DEFAULT_HANDLE); // TODO This should be the actor making the request
-
-        if (!post.id) {
-            ctx.get('logger').info('Invalid Repost - no post id');
-            return;
-        }
-
-        const account = await accountRepository.getBySite(ctx.get('site'));
-        if (account !== null) {
-            const originalPostResult = await postService.getByApId(post.id);
-            if (!isError(originalPostResult)) {
-                const originalPost = getValue(originalPostResult);
-                originalPost.addRepost(account);
-                await postRepository.save(originalPost);
-            }
-        }
-
-        const announce = new Announce({
-            id: announceId,
-            actor: actor,
-            object: post,
-            to: PUBLIC_COLLECTION,
-            cc: apCtx.getFollowersUri(ACTOR_DEFAULT_HANDLE),
-        });
-
-        const announceJson = await announce.toJsonLd();
-
-        // Add announce activity to the database
-        await ctx.get('globaldb').set([announce.id!.href], announceJson);
-        await addToList(ctx.get('db'), ['reposted'], announce.id!.href);
-
-        // Add announce activity to the actor's outbox
-        await addToList(ctx.get('db'), ['outbox'], announce.id!.href);
-
-        // Send the announce activity
-        let attributionActor: Actor | null = null;
-        if (post.attributionId) {
-            attributionActor = await lookupActor(
-                apCtx,
-                post.attributionId.href,
-            );
-        }
-        if (attributionActor) {
-            apCtx.sendActivity(
-                { handle: ACTOR_DEFAULT_HANDLE },
-                attributionActor,
-                announce,
-                {
-                    preferSharedInbox: true,
-                },
-            );
-        }
-
-        apCtx.sendActivity(
-            { handle: ACTOR_DEFAULT_HANDLE },
-            'followers',
-            announce,
-            {
-                preferSharedInbox: true,
-            },
-        );
-
-        return new Response(JSON.stringify(announceJson), {
-            headers: {
-                'Content-Type': 'application/activity+json',
-            },
-            status: 200,
-        });
-    };
-}
-
-export function createDerepostActionHandler(
-    accountRepository: KnexAccountRepository,
-    postService: PostService,
-    postRepository: KnexPostRepository,
-) {
-    return async function derepostAction(
-        ctx: Context<{ Variables: HonoContextVariables }>,
-    ) {
-        const id = ctx.req.param('id');
-        const apCtx = fedify.createContext(ctx.req.raw as Request, {
-            db: ctx.get('db'),
-            globaldb: ctx.get('globaldb'),
-            logger: ctx.get('logger'),
-        });
-
-        const post = await lookupObject(apCtx, id);
-        if (!post) {
-            return new Response(JSON.stringify({ error: 'Post not found' }), {
-                status: 404,
-            });
-        }
-
-        const announceId = apCtx.getObjectUri(Announce, {
-            id: createHash('sha256').update(post.id!.href).digest('hex'),
-        });
-
-        const undoId = apCtx.getObjectUri(Undo, {
-            id: createHash('sha256').update(announceId.href).digest('hex'),
-        });
-
-        const announceToUndoJson = await ctx
-            .get('globaldb')
-            .get([announceId.href]);
-
-        if (!announceToUndoJson) {
-            return new Response(
-                JSON.stringify({ error: 'Repost activity not found' }),
-                {
-                    status: 409,
-                },
-            );
-        }
-
-        const announceToUndo = await Announce.fromJsonLd(announceToUndoJson);
-
-        const actor = await apCtx.getActor(ACTOR_DEFAULT_HANDLE); // TODO This should be the actor making the request
-
-        const idAsUrl = parseURL(id);
-
-        if (!idAsUrl) {
-            return new Response(
-                JSON.stringify({ error: 'ID should be a valid URL' }),
-                {
-                    status: 400,
-                },
-            );
-        }
-
-        const account = await accountRepository.getBySite(ctx.get('site'));
-        const originalPostResult = await postService.getByApId(idAsUrl);
-
-        if (isError(originalPostResult)) {
-            const error = getError(originalPostResult);
-            switch (error) {
-                case 'upstream-error':
-                    ctx.get('logger').info(
-                        'Upstream error fetching post for dereposting',
-                        { postId: idAsUrl.href },
-                    );
-                    break;
-                case 'not-a-post':
-                    ctx.get('logger').info(
-                        'Resource for dereposting is not a post',
-                        { postId: idAsUrl.href },
-                    );
-                    break;
-                case 'missing-author':
-                    ctx.get('logger').info(
-                        'Post for dereposting has missing author',
-                        { postId: idAsUrl.href },
-                    );
-                    break;
-                default:
-                    return exhaustiveCheck(error);
-            }
-        } else {
-            const originalPost = getValue(originalPostResult);
-            originalPost.removeRepost(account);
-            await postRepository.save(originalPost);
-        }
-
-        const undo = new Undo({
-            id: undoId,
-            actor: actor,
-            object: announceToUndo,
-            to: PUBLIC_COLLECTION,
-            cc: apCtx.getFollowersUri(ACTOR_DEFAULT_HANDLE),
-        });
-
-        // Add the undo activity to the database
-        const undoJson = await undo.toJsonLd();
-        await ctx.get('globaldb').set([undo.id!.href], undoJson);
-
-        // Remove announce activity from database
-        await removeFromList(ctx.get('db'), ['reposted'], announceId.href);
-        await ctx.get('globaldb').delete([announceId.href]);
-
-        // Remove announce activity from the actor's outbox
-        await removeFromList(ctx.get('db'), ['outbox'], announceId.href);
-
-        // Send the undo activity
-        let attributionActor: Actor | null = null;
-        if (post.attributionId) {
-            attributionActor = await lookupActor(
-                apCtx,
-                post.attributionId.href,
-            );
-        }
-        if (attributionActor) {
-            apCtx.sendActivity(
-                { handle: ACTOR_DEFAULT_HANDLE },
-                attributionActor,
-                undo,
-                {
-                    preferSharedInbox: true,
-                },
-            );
-        }
-
-        apCtx.sendActivity(
-            { handle: ACTOR_DEFAULT_HANDLE },
-            'followers',
-            undo,
-            {
-                preferSharedInbox: true,
-            },
-        );
-
-        return new Response(JSON.stringify(undoJson), {
-            headers: {
-                'Content-Type': 'application/activity+json',
-            },
-            status: 200,
-        });
-    };
 }
